requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.4
pydantic>=2.4.2
fastapi>=0.104.0
uvicorn>=0.23.2
httpx>=0.25.0
python-multipart>=0.0.6
pymongo>=4.5.0
requests>=2.31.0
numpy>=1.26.0
moviepy>=1.0.3
pillow>=10.1.0
elevenlabs>=0.2.26
python-dotenv>=1.0.0
pedalboard>=0.7.6
whisper>=1.1.10
ffmpeg-python>=0.2.0
librosa>=0.10.1
pydub>=0.25.1
loguru>=0.7.0
pytest>=7.4.0
aiohttp>=3.8.5
tqdm>=4.66.0
jinja2>=3.0.0


